- Removed TuxConfiguration - Moved IDashboardService from interfaces into Services folder
