--- Tuxboard.Core/Configuration/TuxConfiguration.cs
@@ -1,7 +1,0 @@
-﻿namespace Tuxboard.Core.Configuration;
-
-public class TuxConfiguration
-{
-    public const string WidgetDefaultPath = "~/TuxboardFeature/Widgets/";
-    public const string WidgetDefaultFileName = "Default.cshtml";
-}
