+++ Tuxboard.Core/Configuration/TuxboardConfigDto.cs
@@ -1,9 +1,24 @@
 ﻿namespace Tuxboard.Core.Configuration;
 
+/// <summary>
+/// 
+/// </summary>
 public class TuxboardConfigDto
 {
+    /// <summary>
+    /// 
+    /// </summary>
     public bool Moveable { get; set; }
+    /// <summary>
+    /// 
+    /// </summary>
     public bool DeleteWidgets { get; set; }
+    /// <summary>
+    /// 
+    /// </summary>
     public bool UseSettings { get; set; }
+    /// <summary>
+    /// 
+    /// </summary>
     public bool Personalized { get; set; }
 }
