+++ Tuxboard.Core/Data/Configuration/DashboardConfiguration.cs
@@ -6,11 +6,19 @@ using Tuxboard.Core.Domain.Entities;
 
 namespace Tuxboard.Core.Data.Configuration;
 
+/// <summary>
+/// Entity Framework Configuration for <see cref="Dashboard"/> for entity properties and relationships
+/// </summary>
 public class DashboardConfiguration : IEntityTypeConfiguration<Dashboard>
 {
     private readonly TuxboardConfig _config;
     private readonly Action<EntityTypeBuilder<Dashboard>> _seedAction;
 
+    /// <summary>
+    /// 
+    /// </summary>
+    /// <param name="config"></param>
+    /// <param name="seedAction"></param>
     public DashboardConfiguration(TuxboardConfig config,
         Action<EntityTypeBuilder<Dashboard>> seedAction = null)
     {
@@ -18,6 +26,7 @@ public class DashboardConfiguration : IEntityTypeConfiguration<Dashboard>
         _seedAction = seedAction;
     }
 
+    /// <inheritdoc />
     public void Configure(EntityTypeBuilder<Dashboard> builder)
     {
         builder.ToTable("Dashboard", _config.Schema);
