+++ Tuxboard.Core/Data/Configuration/PlanConfiguration.cs
@@ -6,11 +6,19 @@ using Tuxboard.Core.Domain.Entities;
 
 namespace Tuxboard.Core.Data.Configuration;
 
+/// <summary>
+/// Entity Framework Configuration for <see cref="Plan"/> for entity properties and relationships
+/// </summary>
 public class PlanConfiguration : IEntityTypeConfiguration<Plan>
 {
     private readonly TuxboardConfig _config;
     private readonly Action<EntityTypeBuilder<Plan>> _seedAction;
 
+    /// <summary>
+    /// 
+    /// </summary>
+    /// <param name="config"></param>
+    /// <param name="seedAction"></param>
     public PlanConfiguration(TuxboardConfig config,
         Action<EntityTypeBuilder<Plan>> seedAction = null)
     {
@@ -18,6 +26,7 @@ public class PlanConfiguration : IEntityTypeConfiguration<Plan>
         _seedAction = seedAction;
     }
 
+    /// <inheritdoc />
     public void Configure(EntityTypeBuilder<Plan> builder)
     {
         builder.ToTable("Plan", _config.Schema);
