+++ Tuxboard.Core/Infrastructure/Extensions/ServiceCollectionExtensions.cs
@@ -5,7 +5,6 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Tuxboard.Core.Configuration;
 using Tuxboard.Core.Data.Context;
-using Tuxboard.Core.Infrastructure.Interfaces;
 using Tuxboard.Core.Infrastructure.Services;
 
 namespace Tuxboard.Core.Infrastructure;
