+++ Tuxboard.Core/Tuxboard.Core.csproj
@@ -46,4 +46,8 @@
     <Service Include="{508349b6-6b84-4df5-91f0-309beebad82d}" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Infrastructure\Interfaces\" />
+  </ItemGroup>
+
 </Project>
